algotrader/requirements_minimal.txt
python-decouple>=3.8
pandas
numpy
websockets
watchdog
requests
pandas-ta
python-dateutil
pytz
tzdata
MetaTrader5
